Create background_video.js
